+++ packages/@expo/cli/src/install/__tests__/resolveOptions-test.ts
@@ -32,7 +32,10 @@ describe(resolveArgsAsync, () => {
   });
   it(`prevents bad combos`, async () => {
     await expect(resolveArgsAsync(['--npm', '--yarn'])).rejects.toThrow(
-      /Specify at most one of: --npm, --yarn/
+      /Specify at most one of: --npm, --pnpm, --yarn/
+    );
+    await expect(resolveArgsAsync(['--npm', '--pnpm', '--yarn'])).rejects.toThrow(
+      /Specify at most one of: --npm, --pnpm, --yarn/
     );
   });
   it(`allows known values`, async () => {