+++ packages/@expo/cli/src/install/resolveOptions.ts
@@ -3,7 +3,7 @@ import * as PackageManager from '@expo/package-manager';
 import * as Log from '../log';
 import { CommandError } from '../utils/errors';
 
-export type Options = Pick<PackageManager.CreateForProjectOptions, 'npm' | 'yarn'> & {
+export type Options = Pick<PackageManager.CreateForProjectOptions, 'npm' | 'pnpm' | 'yarn'> & {
   /** Check which packages need to be updated, does not install any provided packages. */
   check?: boolean;
   /** Should the dependencies be fixed automatically. */
@@ -14,8 +14,8 @@ function resolveOptions(options: Options): Options {
   if (options.fix && options.check) {
     throw new CommandError('BAD_ARGS', 'Specify at most one of: --check, --fix');
   }
-  if (options.npm && options.yarn) {
-    throw new CommandError('BAD_ARGS', 'Specify at most one of: --npm, --yarn');
+  if ([options.npm, options.pnpm, options.yarn].filter(Boolean).length > 1) {
+    throw new CommandError('BAD_ARGS', 'Specify at most one of: --npm, --pnpm, --yarn');
   }
   return {
     ...options,
@@ -68,7 +68,7 @@ export async function resolveArgsAsync(
 ): Promise<{ variadic: string[]; options: Options; extras: string[] }> {
   const { variadic, extras, flags } = parseVariadicArguments(argv);
 
-  assertUnexpectedObjectKeys(['--check', '--fix', '--npm', '--yarn'], flags);
+  assertUnexpectedObjectKeys(['--check', '--fix', '--npm', '--pnpm', '--yarn'], flags);
 
   return {
     // Variadic arguments like `npx expo install react react-dom` -> ['react', 'react-dom']
@@ -78,6 +78,7 @@ export async function resolveArgsAsync(
       check: !!flags['--check'],
       yarn: !!flags['--yarn'],
       npm: !!flags['--npm'],
+      pnpm: !!flags['--pnpm'],
     }),
     extras,
   };