+++ packages/@expo/cli/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Validate Android SDK configuration before using ([#17259](https://github.com/expo/expo/pull/17259) by [@byCedric](https://github.com/byCedric))
 - Add CI context to telemetry to help determine support on used CI providers ([#17284](https://github.com/expo/expo/pull/17284) by [@byCedric](https://github.com/byCedric))
+- add `--pnpm` option to `install` command.
 
 ### 🐛 Bug fixes